requirements.txt
aiopg==0.16.0
psycopg2-binary==2.8.2
tornado==6.0.2
